Add: vault service
This will enable the Hashicorp Vault as a fully functioning service. Try it out!
Perhaps the next step is to encapsulate holochain into a service on here as
well? I could start working on a PR to upstream nixpkgs if that makes sense

Remove: vault packages

Vault Packages aren't needed at all if we are declaring it as a service. This
should also become true for holochain: it needs to become a service which simply
runs when you declare it.